Se modifica la implementacion para no modificar la version de jets3t que se esta utilizando inicialmente en mu-commander

+++ aether-examples/mu-commander/mu-commander-aether/src/main/java/com/mucommander/file/impl/s3/S3Object.java
@@ -25,10 +25,9 @@ import com.mucommander.io.BufferPool;
 import com.mucommander.io.FileTransferException;
 import com.mucommander.io.RandomAccessInputStream;
 import com.mucommander.io.StreamUtils;
+//import org.jets3t.service.S3Service;
 import org.jets3t.service.S3ServiceException;
-import org.jets3t.service.ServiceException;
-import org.jets3t.service.model.StorageObject;
-import org.jets3t.service.model.StorageOwner;
+import org.jets3t.service.model.S3Owner;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -165,9 +164,9 @@ public class S3Object extends S3File {
         try {
             // Make sure that the directory is empty, abort if not.
             // Note that we must not count the parent directory (this file).
-//            if(service.listObjectsChunked(bucketName, getObjectKey(isDirectory()), "/", 2, null, false).getObjects().length>=2) {
-//                throw new IOException("Directory not empty");
-//            }
+            if(service.listObjectsChunked(bucketName, getObjectKey(isDirectory()), "/", 2, null, false).getObjects().length>=2) {
+                throw new IOException("Directory not empty");
+            }
             service.deleteObject(bucketName, getObjectKey(isDirectory()));
 
             // Update file attributes locally
@@ -175,7 +174,7 @@ public class S3Object extends S3File {
             atts.setDirectory(false);
             atts.setSize(0);
         }
-        catch(ServiceException e) {
+        catch(S3ServiceException e) {
             throw getIOException(e);
         }
     }
@@ -214,8 +213,8 @@ public class S3Object extends S3File {
             destObjectFile.atts.setAttributes(destObject);
             destObjectFile.atts.setExists(true);
         }
-        catch(Exception e) {
-            throw new IOException(e.getMessage());
+        catch(S3ServiceException e) {
+            throw getIOException(e);
         }
     }
 
@@ -231,8 +230,8 @@ public class S3Object extends S3File {
             // add unnecessary billing overhead since it reads the object chunk by chunk, each in a separate GET request.
             return service.getObject(bucketName, getObjectKey(false), null, null, null, null, offset==0?null:offset, null).getDataInputStream();
         }
-        catch(Exception e) {
-            throw new IOException(e.getMessage());
+        catch(S3ServiceException e) {
+            throw getIOException(e);
         }
     }
 
@@ -450,8 +449,8 @@ public class S3Object extends S3File {
                         .getDataInputStream();
                     this.offset = offset;
                 }
-                catch(Exception e) {
-                    throw new IOException(e.getMessage());
+                catch(S3ServiceException e) {
+                    throw getIOException(e);
                 }
             }
         }
@@ -605,33 +604,23 @@ public class S3Object extends S3File {
             updateExpirationDate(); // declare the attributes as 'fresh'
         }
 
-        private void setAttributes(StorageObject storageObject) {
-            setDirectory(storageObject.getKey().endsWith("/"));
-            setSize(storageObject.getContentLength());
-            setDate(storageObject.getLastModifiedDate().getTime());
+        private void setAttributes(org.jets3t.service.model.S3Object object) {
+            setDirectory(object.getKey().endsWith("/"));
+            setSize(object.getContentLength());
+            setDate(object.getLastModifiedDate().getTime());
             setPermissions(DEFAULT_PERMISSIONS);
             // Note: owner is null for common prefix objects
-            StorageOwner owner = storageObject.getOwner();
+            S3Owner owner = object.getOwner();
             setOwner(owner==null?null:owner.getDisplayName());
         }
 
         private void fetchAttributes() throws AuthException {
             try {
-            	if (getObjectKey() != null && !getObjectKey().endsWith("/"))
-            		setAttributes(service.getObjectDetails(bucketName, getObjectKey()));
-            	else { //es un directorio
-                    setDirectory(true);
-                    setSize(0);
-                    setDate(System.currentTimeMillis());
-                    setPermissions(DEFAULT_PERMISSIONS);
-                    // Note: owner is null for common prefix objects
-                    setOwner(null);
-            	}
-            		
+                setAttributes(service.getObjectDetails(bucketName, getObjectKey(), null, null, null, null));
                 // Object does not exist on the server
                 setExists(true);
             }
-            catch(ServiceException e) {
+            catch(S3ServiceException e) {
                 // Object does not exist on the server, or could not be retrieved
                 setExists(false);
